adding the couch package, moving the cache stuff around to handle cache options
