+++ actions/actionctx/get.go
@@ -14,9 +14,9 @@ func GetEvent(ctx context.Context) (events.Event, bool) {
 	return v, ok
 }
 
-// GetStaticDevice returns a static device/true from ctx if there is one, and false if there isn't.
-func GetStaticDevice(ctx context.Context) (statedefinition.StaticDevice, bool) {
-	v, ok := ctx.Value(staticDevice).(statedefinition.StaticDevice)
+// GetStaticDevices returns a static device/true from ctx if there is one, and false if there isn't.
+func GetStaticDevices(ctx context.Context) ([]statedefinition.StaticDevice, bool) {
+	v, ok := ctx.Value(staticDevice).([]statedefinition.StaticDevice)
 	return v, ok
 }
