+++ config/config.go
@@ -30,7 +30,7 @@ const (
 
 //Config .
 type Config struct {
-	Jobs       []JobConfig `json:"jobs"`
+	//	Jobs       []JobConfig `json:"jobs"`
 	Forwarders []Forwarder `json:"forwarders"`
 	Caches     []Cache     `json:"caches"`
 }
