+++ couch/couch.go
@@ -1,0 +1,85 @@
+package couch
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"os"
+	"sync"
+
+	"github.com/byuoitav/common/log"
+	"github.com/byuoitav/common/nerr"
+)
+
+var (
+	username string
+	password string
+	once     sync.Once
+)
+
+func initialize() {
+	log.L.Info("Initializing couch requests")
+	username = os.Getenv("COUCH_USER")
+	password = os.Getenv("COUCH_PASS")
+}
+
+//MakeRequest makes a generic Couch Request
+func MakeRequest(addr, method string, body interface{}) ([]byte, *nerr.E) {
+	once.Do(initialize)
+
+	log.L.Debugf("Making couch request against: %s", addr)
+
+	var reqBody []byte
+	var err error
+
+	// marshal request if not already an array of bytes
+	switch v := body.(type) {
+	case []byte:
+		reqBody = v
+	default:
+		// marshal the request
+		reqBody, err = json.Marshal(v)
+		if err != nil {
+			return []byte{}, nerr.Translate(err)
+		}
+	}
+
+	// create the request
+	req, err := http.NewRequest(method, addr, bytes.NewReader(reqBody))
+	if err != nil {
+		return []byte{}, nerr.Translate(err)
+	}
+
+	// add auth
+	req.SetBasicAuth(username, password)
+
+	// add headers
+	if method == http.MethodPost || method == http.MethodPut {
+		req.Header.Add("content-type", "application/json")
+	}
+
+	client := http.Client{}
+
+	resp, err := client.Do(req)
+	if err != nil {
+		return []byte{}, nerr.Translate(err)
+	}
+	defer resp.Body.Close()
+
+	// read the resp
+	respBody, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return []byte{}, nerr.Translate(err)
+	}
+
+	// check resp code
+	if resp.StatusCode/100 != 2 {
+		msg := fmt.Sprintf("non 200 reponse code received. code: %v, body: %s", resp.StatusCode, respBody)
+		return respBody, nerr.Create(msg, http.StatusText(resp.StatusCode))
+	}
+
+	return respBody, nil
+
+}
