+++ version.txt
@@ -1,1 +1,1 @@
-0.0.22
+0.0.23
